Improve ts documentation
When a user navigates through the code and checking library declaration, the documentation of `then` is misleading.

It can be confusing to know if the initial value is used in the second validator or the result of the first one. To be sure, we need to look at the repository readme.

This commit improve this particular documentation and add an example.